feat: blog page moved to own file and new index page created
